backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
sqlalchemy==2.0.38
psycopg[binary]==3.2.4
alembic==1.14.1
pydantic==2.10.6
pydantic-settings==2.8.0
httpx==0.28.1
python-multipart==0.0.20
pypdf==5.3.1
python-jose[cryptography]==3.3.0
passlib==1.7.4
email-validator==2.2.0
